fix stylesheet import
