+++ src/main/python/jobs.py
@@ -20,6 +20,8 @@
         self.current_row = 0            # param number
 """
 
+from fbs_runtime.application_context.PyQt5 import ApplicationContext
+
 from mainview import MainView
 from uix.util import *
 from uix.config import *
@@ -27,9 +29,365 @@ from uix.popup import Question
 from api import Api
 
 
+
+page_style = """
+
+    #Page_sub_page {
+        background-color: white;
+    }
+    
+    #Page_machine_config {
+        background-color: #D6EFEF;
+    }
+    
+    #Page_resource_planning {
+        background-color: #F4F8F9;
+    }
+    
+    #Page_resource_submission {
+        background-color: #F4F8F9;
+    }
+    
+    #Page_scheme {
+        background-color: #F4F8F9;
+    }
+    
+    #Page_available_resources {
+        background-color: #D6EFEF;
+    }
+    
+    #Page_job_submission {
+        background-color: #F4F8F9;
+    }
+    
+    #Page_section_title {
+        font-family: "Helvetica Neue";
+        font-size: 20px;
+        font-weight: 100;
+        color: #6C7E8E;
+    }
+    
+    #Page_section_title_small {
+        font-family: "Helvetica Neue";
+        font-size: 16px;
+        font-weight: 300;
+        color: #6C7E8E;
+    }
+    
+    #Page_hint {
+        font-family: "Helvetica Neue";
+        font-size: 12px;
+        font-weight: 300;
+        color: red;
+    }
+    
+    #Page_hint_small {
+        font-family: "Helvetica Neue";
+        font-size: 11px;
+        font-weight: 300;
+        color: red;
+    }
+
+    #Page_input_frame {
+        background-color: white;
+    }
+
+    #Page_input_title {
+        font-family: "Helvetica Neue";
+        font-size: 13px;
+        font-weight: 300;
+        color: #6C7E8E;
+    }
+    
+    #Page_button {
+        border: None;
+        background-color: #6C7E8E;
+        height: 30px;
+        width: 120px;
+        font-family: "Helvetica Neue";
+        font-size: 12px;
+        font-weight: 300;
+        color: white;
+    }
+
+    #Page_available_title {
+        font-family: "Helvetica Neue";
+        font-size: 12px;
+        font-weight: 500;
+        color: #6C7E8E;
+    }
+    
+    #Page_available_label {
+        font-family: "Helvetica Neue";
+        font-size: 11px;
+        font-weight: 500;
+        color: #6C7E8E;
+    }
+    
+    #Page_table_workspace {
+        background-color: white;
+    }
+
+    #Page_table_test {
+        background-color: yellow;
+    }
+    
+    #Page_table_workspace_search {
+        border: None;
+        background-color: #F7F7F7;
+        padding: 0 20px;
+        font-family: "Helvetica Neue";
+        font-size: 13px;
+        font-weight: 200;
+        color: black;
+    }
+    
+    #Page_table_workspace_button {
+        border: None;
+        background-color: #6C7E8E;
+        height: 35px;
+        width: 90px;
+        font-family: "Helvetica Neue";
+        font-size: 13px;
+        font-weight: 200;
+        color: white;
+    }
+    
+    #Page_table {
+        border: none;
+        background-color: white;
+        alternate-background-color: #FAFAFA;
+    }
+    
+    #Page_table QHeaderView::section {
+        border: none;
+        background-color: #6C7E8E;
+        height: 35px;
+        font-family: "Helvetica Neue";
+        font-size:13px;
+        font-weight: 100;
+        color: white;
+    }
+"""
+
+page_menu_button_active = f"""
+    border: None;
+    border-bottom: 2px solid {COLOR_01};
+    width: 95px;
+    height: 27px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 200;
+    color: {COLOR_01};
+"""
+
+page_menu_button = """
+    border: None;
+    width: 95px;
+    height: 27px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 100;
+    color: #CBCBCB;
+"""
+
+Page_input_ip_input = """
+    border: None;
+    background-color: #F7F7F7;
+    padding: 0 1px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 200;
+    color: black;
+"""
+
+Page_input_ip_input_disable = f"""
+    border: None;
+    background-color: {COLOR_08};
+    padding: 0 1px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 200;
+    color: {COLOR_09};
+"""
+
+
+Page_machine_config_box = """
+    background-color: white;
+"""
+
+Page_machine_config_box_disable = f"""
+    background-color: {COLOR_08};
+"""
+
+Page_machine_config_label = f"""
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 200;
+    color: {COLOR_01};
+"""
+
+Page_machine_config_label_red = f"""
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 500;
+    color: red;
+"""
+
+Page_machine_config_label_green = f"""
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 400;
+    color: #95E637;
+"""
+
+Page_machine_config_label_disable = f"""
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 400;
+    color: #F2F2F2;
+"""
+
+Page_input_input = f"""
+    border: None;
+    background-color: white;
+    padding: 0 1px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 200;
+    color: black;
+"""
+
+Page_input_input_disable = f"""
+    border: None;
+    background-color: {COLOR_08};
+    padding: 0 1px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 200;
+    color: {COLOR_09};
+"""
+
+Page_evaluate_button = f"""
+    border: None;
+    background-color: {COLOR_01};
+    height: 30px;
+    width: 120px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 300;
+    color: white;   
+"""
+
+Page_evaluate_button_disable = f"""
+    border: None;
+    background-color: {COLOR_08};
+    height: 30px;
+    width: 120px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 300;
+    color: {COLOR_09};
+"""
+
+Page_submission_button = f"""
+    border: none;
+    background-color: white;
+    padding-bottom: 3px;
+    height: 20px;
+    width: 23px;
+    font-size: 20px;
+    color: {COLOR_01};
+"""
+
+Page_submission_button_disable = f"""
+    border: none;
+    background-color: {COLOR_08};
+    padding-bottom: 3px;
+    height: 20px;
+    width: 23px;
+    font-size: 20px;
+    color: {COLOR_09};
+"""
+
+Page_submission_box = f"""
+    background-color: white;
+"""
+
+Page_submission_box_disable = f"""
+    background-color: {COLOR_08};
+"""
+
+Page_submission_label = f"""
+    font-family: "Helvetica Neue";
+    font-size: 13px;
+    font-weight: 300;
+    color: {COLOR_01};
+"""
+
+Page_submission_label_disable = f"""
+    font-family: "Helvetica Neue";
+    font-size: 13px;
+    font-weight: 300;
+    color: {COLOR_09};
+"""
+
+Page_submission_submit = f"""
+    border: None;
+    background-color: {COLOR_01};
+    height: 30px;
+    width: 90px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 300;
+    color: white;
+"""
+
+Page_submission_submit_disable = f"""
+    border: None;
+    background-color: {COLOR_08};
+    height: 30px;
+    width: 90px;
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 300;
+    color: {COLOR_09};
+"""
+
+Page_scheme_box = f"""
+    background-color: #82B9B9;
+"""
+
+Page_scheme_box_disable = f"""
+    background-color: {COLOR_03};
+"""
+
+Page_scheme_button_frame = f"""
+    border-color: {COLOR_01};
+    border-width: 1px;        
+    border-style: solid;
+
+    background-color: white;
+"""
+
+Page_scheme_label = f"""
+    font-family: "Helvetica Neue";
+    font-size: 11px;
+    font-weight: 400;
+    color: white;
+"""
+
+Page_scheme_label_disable = f"""
+    font-family: "Helvetica Neue";
+    font-size: 11px;
+    font-weight: 400;
+    color: {COLOR_01};
+"""
+
+
 class Jobs(MainView):
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, cxt:ApplicationContext, *args, **kwargs):
         super(Jobs, self).__init__(*args, **kwargs)
 
         # variable
