+++ src/main/resources/base/app.qss
@@ -1,0 +1,82 @@
+#App_sidebar {
+    background-color: #FEFEFE;
+    border-right: 1px solid #E7EEF0;
+}
+
+#App_navigation {
+    background-color: white;
+}
+
+#App_main_window {
+    background-color: #F4F8F9;
+}
+
+#App_mask {
+    background-color: rgba(255, 255, 255, 150);
+}
+
+#App_sidebar_title {
+    font-family: "Helvetica Neue";
+    font-size: 22px;
+    font-weight: 500;
+    color: #505A5F;
+}
+
+#App_navigation_button {
+    border: None;
+}
+
+#App_navigation_credit {
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 300;
+    color: #505A5F;
+}
+
+#App_mask_clicked_area {
+    border: none;
+}
+
+#App_account {
+    background-color: white;
+}
+
+#App_title_frame {
+    border: none;
+    border-bottom: 1px solid #D5D5D5;
+}
+
+#App_account_username {
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 500;
+    color: #6C7E8E;
+}
+
+#App_account_credit {
+    font-family: "Helvetica Neue";
+    font-size: 11px;
+    font-weight: 100;
+    color: #505A5F;
+}
+
+#App_account_button {
+    border: none;
+    font-family: "Helvetica Neue";
+    font-size: 10px;
+    font-weight: 200;
+    color: #6C7E8E;
+    text-align: left;
+}
+
+#App_account_logout {
+    border: none;
+    border-radius: 3px;
+    background-color: #6C7E8E;
+    height: 25px;
+    width: 85px;
+    font-family: "Helvetica Neue";
+    font-size: 11px;
+    font-weight: 200;
+    color: white;
+}
