+++ src/main/resources/base/question.qss
@@ -1,0 +1,33 @@
+#Question_window {
+    border: none;
+    background-color: #F4F8F9;
+}
+
+#Question_question {
+    font-family: "Helvetica Neue";
+    font-size: 12px;
+    font-weight: 400;
+    color: #6C7E8E;
+}
+
+#Question_button_confirm {
+    border: none;
+    background-color: #0069D9;
+    height: 22px;
+    width: 65px;
+    font-family: "Helvetica Neue";
+    font-size: 10px;
+    font-weight: 500;
+    color: white;
+}
+
+#Question_button_cancel {
+    border: none;
+    background-color: #B9CCDD;
+    height: 22px;
+    width: 65px;
+    font-family: "Helvetica Neue";
+    font-size: 10px;
+    font-weight: 500;
+    color: white;
+}
